Added FragmentsLab programming assignment
This was a quick lab and I got the hang of it pretty easily. Only issue
is that I cannot get the app to retain its state after configuration
changes, and I've added what I think is appropriate code in both the
Friends and Feed fragments. Might come back to this in the future.